Work on memory usage of es-loader
There is a problem where the bulk* helpers from elasticsearch do not
release the memory associated with processed elements. This means that
the 56GB wikipedia dump ends up consuming all memory.

The streaming xml library also requires that the elements be explicitly
released in order to allow them to be reclaimed.

At this point the es-loader script runs using a few hundred mbs of
memory.